refactor(model): split GatedDeltaNet into ssm/gdn package
Restructure megatron/core/ssm/gated_delta_net.py into a package:
- gdn/gdn_common.py: shared _GDNBase (init/forward/checkpoint skeleton with
  variant hooks), CP utilities, and the torch-native deterministic reference
- gdn/gated_delta_net.py: the GDN variant (in_proj sizing, split tables, gate
  parameters, per-head decay/beta computation, kernel selection)
- megatron/core/ssm/gated_delta_net.py stays as a backward-compatible re-export
  shim; internal imports move to the new package

Also accept (and ignore) the cp_comm_type kwarg that TransformerLayer forwards
to its attention module, fixing GDN training with context parallelism (the CLI
defaults --cp-comm-type to p2p, which previously crashed GDN layer builds).

No functional change otherwise.
