+++ megatron/core/ssm/gdn/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2026, NVIDIA CORPORATION. All rights reserved.
+
+"""Gated Delta Net (GDN) family of layers."""
+
+from megatron.core.ssm.gdn.gated_delta_net import GatedDeltaNet
+from megatron.core.ssm.gdn.gdn_common import GatedDeltaNetSubmodules
