+++ tests/unit_tests/models/test_experimental_attention_variant_module_specs.py
@@ -230,7 +230,7 @@ class TestGetGatedDeltaNetModuleSpec:
         from megatron.core.models.gpt.experimental_attention_variant_module_specs import (
             get_gated_delta_net_module_spec,
         )
-        from megatron.core.ssm.gated_delta_net import GatedDeltaNet
+        from megatron.core.ssm.gdn import GatedDeltaNet
 
         backend = _make_backend()
         cfg = _make_config(normalization="RMSNorm")
