+++ tests/unit_tests/ssm/test_gated_delta_net.py
@@ -18,8 +18,8 @@ from megatron.core.models.gpt.experimental_attention_variant_module_specs import
 )
 from megatron.core.models.gpt.gpt_model import GPTModel
 from megatron.core.process_groups_config import ProcessGroupCollection
-from megatron.core.ssm.gated_delta_net import (
-    GatedDeltaNet,
+from megatron.core.ssm.gdn import GatedDeltaNet
+from megatron.core.ssm.gdn.gdn_common import (
     _build_head_perm_for_split_sections,
     _build_thd_cp_a2a_perm,
     tensor_a2a_cp2hp,
@@ -247,6 +247,42 @@ class TestGatedDeltaNet:
                 rec_grads[name], base_grads[name]
             ), f"Grad not identical for {name} ({rank=})"
 
+    def test_module_construction(self):
+        gdn = self.gdn
+        assert gdn.in_proj_dim == 2 * gdn.qk_dim + 2 * gdn.v_dim + 2 * gdn.num_value_heads
+        assert gdn.A_log.shape == (gdn.num_value_heads // self.tp_size,)
+        assert gdn.dt_bias.shape == (gdn.num_value_heads // self.tp_size,)
+
+    def test_gpu_forward_backward(self):
+        gdn = self.gdn
+
+        micro_batch_size = 2
+        seq_length = 64
+        hidden_states = torch.rand(
+            (seq_length // self.sp_size // self.cp_size, micro_batch_size, gdn.config.hidden_size),
+            device=torch.cuda.current_device(),
+            dtype=torch.bfloat16,
+            requires_grad=True,
+        )
+
+        output, bias = gdn(hidden_states, None)
+
+        assert output.shape == (
+            seq_length // self.sp_size // self.cp_size,
+            micro_batch_size,
+            gdn.config.hidden_size,
+        ), f"Output shape mismatch ({output.shape=})"
+        assert (
+            output.dtype == hidden_states.dtype
+        ), f"Output dtype {output.dtype=} mismatch with {hidden_states.dtype=}"
+
+        output.float().sum().backward()
+        assert hidden_states.grad is not None
+        assert torch.isfinite(hidden_states.grad).all(), "Non-finite input grad"
+        for name, param in gdn.named_parameters():
+            if param.grad is not None:
+                assert torch.isfinite(param.grad).all(), f"Non-finite grad for {name}"
+
     def test_jit_compiled_helpers(self):
         import torch._dynamo
 
@@ -254,62 +290,52 @@ class TestGatedDeltaNet:
         batch = 2
         seq_len = 16
 
+        device = torch.cuda.current_device()
         num_v_heads_local = gdn.num_value_heads // gdn.tp_size // gdn.cp_size
+        num_k_heads_local = gdn.num_key_heads // gdn.tp_size // gdn.cp_size
+        qk_dim_local = gdn.qk_dim_local_tp // gdn.cp_size
+        v_dim_local = gdn.v_dim_local_tp // gdn.cp_size
 
-        qkv_last_dim = (2 * gdn.qk_dim_local_tp + gdn.v_dim_local_tp) // gdn.cp_size
         qkv = torch.randn(
-            batch, seq_len, qkv_last_dim, device=torch.cuda.current_device(), dtype=torch.bfloat16
+            batch, seq_len, 2 * qk_dim_local + v_dim_local, device=device, dtype=torch.bfloat16
         )
         gate = torch.randn(
             batch,
             seq_len,
             num_v_heads_local,
             gdn.value_head_dim,
-            device=torch.cuda.current_device(),
+            device=device,
             dtype=torch.bfloat16,
         )
-        beta = torch.randn(
-            batch,
-            seq_len,
-            num_v_heads_local,
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
-        alpha = torch.randn(
-            batch,
-            seq_len,
-            num_v_heads_local,
-            device=torch.cuda.current_device(),
-            dtype=torch.bfloat16,
-        )
+        gate_feats = (
+            torch.randn(batch, seq_len, num_v_heads_local, device=device, dtype=torch.bfloat16),
+            torch.randn(batch, seq_len, num_v_heads_local, device=device, dtype=torch.bfloat16),
+        )  # beta, alpha
+        A_log_mock = torch.randn(num_v_heads_local, device=device, dtype=torch.bfloat16)
+        dt_bias_mock = torch.randn(num_v_heads_local, device=device, dtype=torch.bfloat16)
+        expected_keys = {"q", "k", "v", "g", "beta"}
 
         # Disable dynamo so coverage.py can trace through the method bodies,
         # which are normally wrapped by @jit_fuser (torch.compile).
         with torch._dynamo.config.patch(disable=True):
-            query, key, value, gate_out, beta_out, alpha_out = (
-                gdn._prepare_qkv_for_gated_delta_rule(qkv, gate, beta, alpha, batch, seq_len)
+            kernel_inputs = gdn._prepare_input_for_gated_delta_rule(
+                qkv, gate, gate_feats, A_log_mock, dt_bias_mock, batch, seq_len
             )
+            gate_out = kernel_inputs.pop("gate")
 
+        assert set(kernel_inputs.keys()) == expected_keys
+        query, key, value, g = (kernel_inputs[name] for name in ("q", "k", "v", "g"))
         assert query.shape == (batch, seq_len, num_v_heads_local, gdn.key_head_dim)
         assert key.shape == (batch, seq_len, num_v_heads_local, gdn.key_head_dim)
         assert value.shape == (batch, seq_len, num_v_heads_local, gdn.value_head_dim)
-        assert query.is_contiguous()
-        assert key.is_contiguous()
-        assert value.is_contiguous()
-
-        A_log_mock = torch.randn(
-            num_v_heads_local, device=torch.cuda.current_device(), dtype=torch.bfloat16
-        )
-        dt_bias_mock = torch.randn(
-            num_v_heads_local, device=torch.cuda.current_device(), dtype=torch.bfloat16
-        )
-
-        with torch._dynamo.config.patch(disable=True):
-            g, beta_sig = gdn._compute_g_and_beta(A_log_mock, dt_bias_mock, alpha, beta)
-
+        # The log-decay must be in fp32
         assert g.dtype == torch.float32
-        assert g.shape == alpha.shape
-        assert beta_sig.shape == beta.shape
+        for t in (query, key, value, gate_out, g):
+            assert t.is_contiguous()
+
+        # Per-head decay and write strength beta
+        assert g.shape == (batch, seq_len, num_v_heads_local)
+        assert kernel_inputs["beta"].shape == (batch, seq_len, num_v_heads_local)
 
     def test_gpu_forward_thd_correctness(self):
         if self.sp_size > 1:
